+++ CHANGELOG.rst
@@ -64,7 +64,7 @@ Bug Fixes
 
 * (Python) Forward reference annotations are no longer rendered as strings.
 * (Python) autoapifunction directive no longer documents async functions as
-    a normal function.
+  a normal function.
 * (Python) Fixed unicode decode errors in some Python 3 situations.